requirements.txt
selenium
numpy
pandas
openpyxl
python-dotenv
pyyaml
webdriver_manager